% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{getDefaultColours}
\alias{getDefaultColours}
\title{Extract CSV of default phyla colours for ncbi or gtdb}
\usage{
getDefaultColours(database_name, save_name)
}
\arguments{
\item{database_name}{(string) the name of the database you would like default phyla colours for (ncbi or gtdb)}

\item{save_name}{(string) the full path to where the CSV should be saved}
}
\value{
Nothing returned, but a CSV is saved
}
\description{
Extract CSV of default phyla colours for ncbi or gtdb
}
