% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{getPhylaCSV}
\alias{getPhylaCSV}
\title{Extract CSV of all phyla names and an empty column for Colour (ncbi or gtdb)}
\usage{
getPhylaCSV(database_name, database_version, save_name)
}
\arguments{
\item{database_name}{(string) the name of the database you would like phyla from (ncbi or gtdb)}

\item{database_version}{(string or list) string: the release version, or empty for the latest version, list: files for your database (see https://github.com/pirovc/multitax for more information)}

\item{save_name}{(string) the full path to where the CSV should be saved}
}
\value{
Nothing returned, but a CSV is saved
}
\description{
Extract CSV of all phyla names and an empty column for Colour (ncbi or gtdb)
}
