% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{getRequestedColours}
\alias{getRequestedColours}
\title{Generate dataframe and optional CSV of colours for requested taxa (global mode or dataset-specific mode)}
\usage{
getRequestedColours(
  database_name,
  database_version,
  requested_taxa,
  name_save_colours,
  name_provided_colours = "",
  mode = "global"
)
}
\arguments{
\item{database_name}{(string) the name of the database you would like all phyla for (ncbi or gtdb)}

\item{database_version}{(string or list) string: the release version, or empty for the latest version, list: files for your database (see https://github.com/pirovc/multitax for more information)}

\item{requested_taxa}{(string or DataFrame): path to a CSV containing column 'Taxon' that is the scientific name of all taxa you need a colour for OR
a dataframe of the same format}

\item{name_save_colours}{(string): name to save requested colours (see OUTPUT) as CSV. If do not want to save CSV, use ""}

\item{name_provided_colours}{(string): OPTIONAL: A path to a CSV of parent colours (at least phyla needed) in format of 2 columns:
\itemize{
\item 'Taxon' contains scientific name of phyla (and additional parent colours at lower taxonomic levels if requested)
\item 'Colour' contains hex code (with leading "#") of the corresponding taxon
}}

\item{mode}{(string): OPTIONAL: can be global (always generate the same colour for the same taxon) or "dataset-specific" (generate colours such that all
taxa in the same phyla are maximally different). Default is "global"}
}
\value{
A dataframe of 2 columns: 1) 'Taxon' with values from name_requested_colours and 2)'Colour' the corresponding hex colour code
}
\description{
Generate dataframe and optional CSV of colours for requested taxa (global mode or dataset-specific mode)
}
